Remove non-necessary fixed options from gpg calls
These options can easily defined in the gnupg option file at `~/.gnupg/gpg.conf` .
Setting them on the command line forces to use particular crypto settings, which can not be overwritten in the option file.
This makes nautilus-scripts unusable, if the options don't fit to the users needs.
So it's better to use gnupg's defaults instead.