rpm-ostree/remove: --apply-live
the --apply-live subcommand adds unnecessary complexity. The removal is due to: in my testing I interrupted rpm-ostree while applying the live commit and was left with the inability to apply the overlayed packages and an OS stuck in a transient state. if you re-attempt, the packages are still layered the apply-live is broken.